requirements.txt
# DMX Bridge Dependencies
# Python 3.7+ required

# sACN (E1.31) protocol support
sacn==1.9.0

# Serial communication with Arduino
pyserial==3.5

# Art-Net protocol support
stupidArtnet==1.3.0
